bird_reconf/requirements.txt
cffi==1.16.0
cryptography==41.0.7
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyYAML==6.0.1
